+++ All.ui/miniprogram/app.js
@@ -19,7 +19,8 @@ App({
 
 
     this.globalData = {
-      statusBarHeight: 0
+      statusBarHeight: 0,
+      DEFAULT_IMG: 'https://profile.csdnimg.cn/E/5/2/1_qq_43297527',
     }
   }
 })